+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 authclient extension Change Log
 
 - Enh #187: URL endpoints for `authUrl` and `tokenUrl` for `yii\authclient\clients\LinkedIn` updated (Felli)
 - Enh #195: `yii\authclient\AuthAction` refactored to use `yii\web\Application::$request` for request data access (klimov-paul)
+- Enh #196: Added `yii\authclient\AuthAction::$cancelCallback` allowing custom handling for authentication cancelation (terales, klimov-paul)
 
 
 2.1.4 November 03, 2017